增加一部分删改方法

+++ src/main/java/com/petplatform/controllor/form_task_delete.java
@@ -1,5 +1,7 @@
 package com.petplatform.controllor;
+import com.petplatform.POJO.Stuff;
 import com.petplatform.POJO.Task;
+import com.petplatform.dao.StuffDao;
 import com.petplatform.dao.TaskDao;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.*;
@@ -13,25 +15,22 @@ import java.lang.reflect.InvocationTargetException;
 import java.util.Date;
 import java.util.Map;
 
-@WebServlet(name = "form_task_delete", value = "/form_task_delete")
+@WebServlet(name = "form_task_delete", value = "/form_task_delete/*")
 public class form_task_delete extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
-        Map<String, String[]> parameterMap = request.getParameterMap();
-
-        Task task = new Task();
-
-        try {
-            BeanUtils.populate(task, parameterMap);
-        } catch (IllegalAccessException | InvocationTargetException e) {
-            throw new RuntimeException(e);
+        String pathInfo = request.getPathInfo();
+        if (pathInfo != null && pathInfo.startsWith("/")) {
+            pathInfo = pathInfo.substring(1); // 移除开头的斜杠
+            String[] parts = pathInfo.split("/");
+            if (parts.length > 0) {
+                String taskNumber = parts[0];
+                Task task = new Task();
+                task.setTaskNumber(Integer.parseInt(taskNumber));
+                TaskDao.deleteTask(task);
+            }
         }
-
-        TaskDao.deleteTask(task);
-        // 重定向到成功页面或者主页面
-        response.sendRedirect(request.getContextPath() + "/admin_html/6_ad_task.html");
 
     }
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
